requirements.txt
streamlit==1.37.1
openai==1.40.0
python-dotenv==1.0.1
pydantic==2.8.2
phonenumbers==8.13.45